1. unit test for new file output method 2. minor changes for re-set calculator method paramaters 3. remove unused library and class 4. rename few class and method name

+++ CodingExercise.IncomeTaxCalculator/Interface/ITaxIncomeCalculator.cs
@@ -11,8 +11,8 @@ namespace CodingExercise.IncomeTaxCalculator.Interface
     /// </summary>
     public interface ITaxIncomeCalculator
     {
-        int CalculateIncomeTax(int annualSalary, TaxBracket taxBracket);
+        int CalculateIncomeTax(decimal annualSalary, TaxBracket taxBracket);
 
-        TaxBracket GetTaxBracket(int salary);
+        TaxBracket GetTaxBracket(decimal salary);
     }
 }
